Add files via upload
In this project, we created an interactive game in the Racket programming language called "Zuul", this is a spin-off of a project students create in the CSCI 225 class which is originally a text-based game written in the Java Programming language. The object of the original java version of Zuul is to teach about java classes, inheritance, interfaces, enums, and the MVC design pattern. In a functional language such as Racket, this project requires a completely different approach which for me was quite fun to code. All of the code was written in the Dr. Racket terminal.