Add files via upload
Implemented a post-script notation calculator in C (ex. + 1 1 = 2). For the implementation, we used a stack to store our values and used a simple switch statement to determine our logic. We started with a simple calculator at stage 1 that could add, subtract, multiply, divide, and perform a modulo calculation and then added support for a single variable, M, in stage 2. In stage 3, we added support for variables A-J.